requirements.txt
python-telegram-bot==21.0.1
PyYAML>=6.0
